requirements.txt
celery==5.3.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.1
Django==3.2.12
django-celery-beat==2.5.0
django-filter==23.2
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
drf-yasg==1.21.7
httplib2==0.20.2
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.3.0
pyparsing==2.4.7
python-crontab==3.0.0
python-dotenv==1.0.0
redis==5.0.0
requests==2.31.0
stripe==6.5.0
gunicorn
django-extensions

